This uses a btree for the object cache for 2 reasons.
1. it is faster than the builtin map
2. it is deterministic, while the builtin map isn't.


Benchmarks on master

```
BenchmarkPreprocess-8                	    1952	    736169 ns/op
BenchmarkLoopyMain-8                 	       1	5493570400 ns/op
BenchmarkMapSet-8                    	85774654	        12.16 ns/op
BenchmarkMapCreateSet-8              	92109302	        14.16 ns/op
BenchmarkMapCreateSetString-8        	71222534	        18.07 ns/op
BenchmarkSliceIterate10-8            	100000000	        12.47 ns/op
BenchmarkStructStack-8                  1000000000	         1.083 ns/op
BenchmarkStructGC-8                  	1000000000	         1.070 ns/op
BenchmarkTypeAssertionMethodCall/Int32().Int32()_(no_interface)-8         	1000000000	         0.3046 ns/op
BenchmarkTypeAssertionMethodCall/BenchValue().Int32()_(interface_method)-8         	778061769	         1.977 ns/op
BenchmarkTypeAssertionMethodCall/v.(Int32).Int32()_(concrete_assert_method)-8      	1000000000	         0.5902 ns/op
BenchmarkTypeAssertionMethodCall/case_v.(Int32).Int32()_(type_switch_concrete_assert_method)-8         	1000000000	         0.9090 ns/op
BenchmarkTypeAssertionMethodCall/MyStruct{Value:Int32(i)}_(struct_interface_field_init)-8              	1000000000	         0.3042 ns/op
BenchmarkTypeAssertionMethodCall/v.(BenchValue).Int32()_(interface_assert_method)-8                    	148655521	         8.056 ns/op
BenchmarkTypeSwitchOrCreate/type-switch-8                                                              	558304	      2169 ns/op
BenchmarkTypeSwitchOrCreate/super-struct-8                                                             	760971	      1547 ns/op
BenchmarkReflectAddInt64-8                                                                             	565599928	         2.398 ns/op
BenchmarkArrayEquality/ArrayEquality[1]-8                                                              	1000000000	         0.7473 ns/op
BenchmarkArrayEquality/ArrayEquality[8]-8                                                              	1000000000	         1.062 ns/op
BenchmarkArrayEquality/ArrayEquality[16]-8                                                             	639341563	         1.916 ns/op
BenchmarkArrayEquality/ArrayEquality[20]-8                                                             	128369622	         9.407 ns/op
BenchmarkArrayEquality/ArrayEquality[32]-8                                                             	148411623	         7.513 ns/op
BenchmarkArrayEquality/ArrayEquality[64]-8                                                             	100638350	        11.75 ns/op
BenchmarkArrayEquality/ArrayEquality[256]-8                                                            	56279106	        22.41 ns/op
```

Benchmarks with a btree

```
BenchmarkPreprocess-8                	    2619	    488234 ns/op
BenchmarkLoopyMain-8                 	       1	3521497000 ns/op
BenchmarkMapSet-8                    	100000000	        11.94 ns/op
BenchmarkMapCreateSet-8              	88771840	        14.17 ns/op
BenchmarkMapCreateSetString-8        	69721231	        17.73 ns/op
BenchmarkSliceIterate10-8            	94301545	        11.86 ns/op
BenchmarkStructStack-8                1000000000	         1.098 ns/op
BenchmarkStructGC-8                  	1000000000	         1.069 ns/op
BenchmarkTypeAssertionMethodCall/Int32().Int32()_(no_interface)-8         	1000000000	         0.3100 ns/op
BenchmarkTypeAssertionMethodCall/BenchValue().Int32()_(interface_method)-8         	796941338	         1.512 ns/op
BenchmarkTypeAssertionMethodCall/v.(Int32).Int32()_(concrete_assert_method)-8      	1000000000	         0.3149 ns/op
BenchmarkTypeAssertionMethodCall/case_v.(Int32).Int32()_(type_switch_concrete_assert_method)-8         	1000000000	         0.9127 ns/op
BenchmarkTypeAssertionMethodCall/MyStruct{Value:Int32(i)}_(struct_interface_field_init)-8              	1000000000	         0.3060 ns/op
BenchmarkTypeAssertionMethodCall/v.(BenchValue).Int32()_(interface_assert_method)-8                    	163830598	         7.254 ns/op
BenchmarkTypeSwitchOrCreate/type-switch-8                                                              	559197	      2173 ns/op
BenchmarkTypeSwitchOrCreate/super-struct-8                                                             	757675	      1530 ns/op
BenchmarkReflectAddInt64-8                                                                             	701443453	         1.808 ns/op
BenchmarkArrayEquality/ArrayEquality[1]-8                                                              	1000000000	         0.3253 ns/op
BenchmarkArrayEquality/ArrayEquality[8]-8                                                              	1000000000	         1.173 ns/op
BenchmarkArrayEquality/ArrayEquality[16]-8                                                             	625828244	         1.817 ns/op
BenchmarkArrayEquality/ArrayEquality[20]-8                                                             	154459993	         7.691 ns/op
BenchmarkArrayEquality/ArrayEquality[32]-8                                                             	185210908	         6.208 ns/op
BenchmarkArrayEquality/ArrayEquality[64]-8                                                             	98414702	        11.93 ns/op
BenchmarkArrayEquality/ArrayEquality[256]-8                                                            	52833178	        23.72 ns/op
```