@peter7891 2023-03-24
> not a major issue, but try to follow the [Contributing guide](https://github.com/gnolang/gno/blob/master/CONTRIBUTING.md#git-commit-messages) on git commit message styles (these are mostly necessary just for PR titles), using [conventional commits](https://www.conventionalcommits.org/en/v1.0.0/). while we're not using any automated tools yet, it does provide a way to categorise changes which is rapidly becoming very common in a lot of open source projects :)

The commit messages don't matter because when i do the merge, i squash them and write a meaningful message there.