@peter7891 2023-03-24
We should probably use this to replace other uses of the builtin map but let's first see how this goes.