@moul 2023-03-27
> The commit messages don't matter because when i do the merge, i squash them and write a meaningful message there

That is accurate, however it is also a sane practice. Please attempt to do it to some degree, as you mentioned the main focus is on the PR as a whole and verifying it at the end.